Adding support for keeping parametrized, the ms/ls-bit of the field. (note: only ms-bit tried and works. Single field in register, ms/ls-bits are [PARAM:0]. nothing else tried).
Note: ms-bit expression must be comprised only of the single parameter
name. No arithmatic or other operands currently supported. (easy enough
to add them later I think..)

Also, Fixing a conceptual bug: Previously if --keep_params was provided
on the command line, *all* parameters were kept unresovled and kept
symbolic.
However, we probably want to do this *selectively* only for the
parameters specified in --keep_params.
This understanding came to me late in the process.
For symbolic register array size + reset value, I didn't do this; if
--keep_params is found on the cmdline, *all* register array dimensions
and field reset values which are parameters, are kept symbolic,
regardless if they appear in -keep_params or not.

But, for the Field MSbit, I now added this support. Only Field MSbits
which are parameters and specified in -keep_params will be kept
symbolic.

So, need to pass 'options' (cmdline options) to the compiler, since the
compiler/elabortor needs to selectively keep symbolic only those
Parameters the use wishes to keep symbolic.